requirements.txt
# Core dependencies
pytesseract>=0.3.13
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Cloud OCR APIs
google-cloud-vision>=3.4.0

# Optional dependencies for enhanced functionality
matplotlib>=3.7.0
scikit-image>=0.21.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
